membuat fitur rental(tambah pesanan)

+++ database/migrations/2026_04_17_125038_create_payments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->foreignId('rental_id')->constrained()->onDelete('cascade');
-            $table->decimal('amount', 8, 2);
+            $table->unsignedInteger('amount');
             $table->string('payment_method');
             $table->string('proof_of_payment');
             $table->enum('status', ['pending', 'verified', 'rejected'])->default('pending');
